Use a computed property for sorting dropdown items.
Create a sortValue computed property for sorting instead of sorting on
the title. Users can provide a sortVal property in the VM that will be
used to order the dropdown items, this way items can be ordered by
something like a date string (YYYYMMDD) while using a more human-
friendly title. If no sortVal property is specified, sortValue falls
back on the title.