Guard against missing indicators.
Wrap the fetching of numerators and denominators in try-catch blocks to
ensure we don't access properties on undefined values. If a value is
missing, it is treated as a 0. For numerators, this means the ratio
value will be 0; for denominators, this means the ratio value will be
Infinity (which is typically treated as undefined by charts).