Fix the sorting of dropdown items
https://seedscientific.atlassian.net/browse/POLIO-203

Remove the sortValue computed property because a computed property on
the dropdown VM can't be used to retrieve the property on a child VM for
the orderBy filter. Instead, use a sortBy property that specifies which
property on the item object to use for sorting, the default is 'title'.
This still allows for custom sort properties on the items.