Remove 90 degree rotation transform
If a transform is applied to the disclosure triangles in the dropdown,
the triangles remain visible even with the menu hidden. This might be a
bug in webkit. To avoid this problem, switch from an fa-caret-right to
an fa-caret-down class based on the open state instead of using the
fa-rotate-90 class to turn the right caret.