Fix selectedItem computation on dropdown component
Vue components don't appear to be able to access the computed properties
of child components, so implementing a selectedItems property on the
dropdown items that traverses the children of that item doesn't work to
generate the list of selected items. Instead, implement a breadth-first
search on the dropdown component itself.