Fix data fetching for choropleth maps
https://seedscientific.atlassian.net/browse/POLIO-281

Specify the start and end date as the start and end of the current campaign
when fetching data for the choropleth map. The choropleth can only display
data for one campaign, so fetching historical values is useless and resulted
in us displaying data for a mixture of different time periods.