Append y values to labels
Naive solution to https://seedscientific.atlassian.net/browse/POLIO-348 is to
just append the formatted value to the series name in the label data's text
property. This works fine for cases where the series label is a word, but
looks weird when the series label is numeric, like the value (e.g. year in the
polio cases chart).