Change the default view to totals, not relative % in the NCO dashboard
I think the total counts are a better initial display because it shows
not only the distribution of the reasons, but also the context of how
many children are included in each region. For example, 67% non-
compliance in Plateau is disturbing until you realize that only 3
children were missed due to non-compliance.