Calculate the menu position relative to $el
Use the containing element (this.$el) to calculate the offset for the
menu, not the menu itself, since the menu's offset changes based on its
previous orientation. This is why the menu would move from being
oriented to the right, to being centered, to being oriented to the left
on subsequent openings.

https://seedscientific.atlassian.net/browse/POLIO-550