Require already transformed series for stacked columns
Create the stacked series objects before passing them to the stacked-
column component instead of passing a flat array of datapoints. This
way we don't rely on using the automatic loading supplied by the
with-indicator mixin for generating series labels, since the immunity
gap data isn't loaded using the with-indicator mixin, and we eventually
want to be rid of that mixin altogether.

* Stack the immunity gap series in the management dashboard component
* Delete the series computed property on the stacked-column component
* Add a series property to the stacked-column VM
* Remove the with-indicator mixin from the stacked-column component