Convert the access plan pie chart to a donut chart
This is supposed to be a donut chart and display the percentage in the
center. Adding the label will take some more time, but for now at least
we can punch a whole in the chart so that it's the correct type of
chart and so that it carries less visual weight on the page.