Fix vertical positioning of labels (again)
The original bounds were being calculated in a different coordinate
space (probably relative to a different parent) than the y-values of the
data objects, resulting in an unnecessary vertical shift when
re-centering labels at the end of the splay function. Update the
findBounds function so that it calculates the bounds based on the data
objects instead of a bounding box and use it for both the original and
the new bounds to guarantee that the calculated delta is accurate.

Also strip off some of the padding the labels were picking up because of
their class name.

https://seedscientific.atlassian.net/browse/POLIO-657