Fix initial display of column headers by porting to React
There was a bug causing the column labels of the heatmap not to be
passed from the district component to the heatmap component. It
seemed to be related to the last change to the headers computed
property not being propagated to the heatmap's headers property.
I avoided the issue entirely by moving the heatmap from Vue to
React (the first of many charts to be ported) so that we don't
rely on opaque two-way bindings as much and instead explicitly
tell charts when to render.