Fix a bug with the time range in column charts
Ensure that the timeRadioValue is non-negative when applying a chart
definition. The default timeValue to search for is allTime, which is not
an option for column charts, and will cause _.findIndex to return -1
which causes errors later.

Removed the workaround from getLower for negative timeRadioValue now
that the root problem is solved.